Implement set 4 and init set 5
Completed set 4 practical, an HTTP client that can send requests
to servers in a similar manner to telnet.

Begun implementing set 5, an HTTP server that accepts HEAD and GET
requests to the root document. At this commit it accepts the requests
and on a GET request will just respond with a single line of text.